moving svg folder to src\img
