--- src/components/Svg/util/getExternalLinkProps.ts
@@ -1,6 +1,0 @@
-const getExternalLinkProps = (): { target: string; rel: string } => ({
-  target: "_blank",
-  rel: "noreferrer noopener",
-});
-
-export default getExternalLinkProps;
