--- src/components/Svg/util/isTouchDevice.ts
@@ -1,11 +1,0 @@
-const isTouchDevice = (): boolean => {
-  return (
-    typeof window !== "undefined" &&
-    ("ontouchstart" in window ||
-      navigator.maxTouchPoints > 0 ||
-      // @ts-ignore
-      navigator.msMaxTouchPoints > 0)
-  );
-};
-
-export default isTouchDevice;
