+++ src/img/Svg/styles.tsx
