+++ src/img/Svg/theme/base.ts
