+++ src/img/Svg/theme/colors.ts
