+++ src/img/Svg/theme/index.ts
