+++ src/img/Svg/theme/types.ts
