+++ src/img/Svg/types.ts
