+++ src/img/Svg/util/getThemeValue.ts
